+++ packages/provider-sdk/src/tools/registry-errors/index.ts
@@ -1,0 +1,3 @@
+export { DuplicateToolError } from "./DuplicateToolError.js";
+export { ToolNotFoundError } from "./ToolNotFoundError.js";
+export { ToolRegistryError } from "./ToolRegistryError.js";